+++ src/assembly/index.ts
@@ -1,6 +1,9 @@
+export * from "./compat";
 export * from "./queryvars";
 export * from "./inference";
-export * from "./connection";
+
+import * as graphql from "./graphql";
+export { graphql };
 
 import * as http from "./http";
 export { http };