+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Remove `@embedder` decorator [#112](https://github.com/hypermodeAI/functions-as/pull/112)
 - Update examples [#113](https://github.com/hypermodeAI/functions-as/pull/113)
+- Misc updates [#114](https://github.com/hypermodeAI/functions-as/pull/114)
+  - `connection.invokeGraphqlApi` is now `graphql.execute`
+  - `connection` and `inference` classes are now marked deprecated
+  - `graphql` and `classification` examples are updated
 
 ## 2024-06-21 - Version 0.9.0